+++ ChestsAnywhere/release-notes.md
@@ -5,7 +5,7 @@
 * Added Automate options to block input or output for a chest.
 * Added owner's name to cabin categories.
 * Separated categories for locations with the same name.
-* Improved translations. Thanks to kelvindules (updated Portuguese) and kice (updated Chinese)!
+* Improved translations. Thanks to kelvindules (updated Portuguese), kice (updated Chinese), and TheOzonO3 (updated Russian)!
 
 ## 1.15.1
 * Added option to disable player inventory organise button.